:sparkles: feat : Add grade score #S09P12A308-132

+++ backend/airlingo/src/main/java/com/ssafy/airlingo/domain/language/repository/GradeRepository.java
@@ -1,12 +1,9 @@
 package com.ssafy.airlingo.domain.language.repository;
 
 import com.ssafy.airlingo.domain.language.entity.Grade;
-import com.ssafy.airlingo.domain.language.entity.Record;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface GradeRepository extends JpaRepository<Grade, Integer> {
+public interface GradeRepository extends JpaRepository<Grade, Long> {
 	Grade findByGradeName(String gradeName);
 }
